refacto: create view_api
